requirements.txt
OpenBCI
Avatar
djitellopy
opencv-python
brainflow
numpy
pandas
matplotlib
scikit-learn
scikit-learn-intelex
scipy
delta-spark
pysftp
PySimpleGUI==4.60.5
flask
tensorflow
tensorflow_decision_forests
requests
flask-socketio